Test for all example numbers provided by Google
