+++ PhoneNumberKitTests/PhoneNumberKitTests.swift
@@ -249,292 +249,5 @@ class PhoneNumberKitTests: XCTestCase {
         let phoneNumberKit = PhoneNumberKit()
         XCTAssertEqual(phoneNumberKit.countriesForCode(424242)?.count, 0)
     }
-    
-    func testUSNumber() {
-        do {
-            let phoneNumber1 = try PhoneNumber(rawNumber: "650 253 0000")
-            XCTAssertNotNil(phoneNumber1)
-            let phoneNumber2 = try PhoneNumber(rawNumber: "+1 650 253 0000")
-            XCTAssertNotNil(phoneNumber2)
-            let phoneNumber3 = try PhoneNumber(rawNumber: "800 253 0000")
-            XCTAssertNotNil(phoneNumber3)
-            let phoneNumber4 = try PhoneNumber(rawNumber: "+1 800 253 0000")
-            XCTAssertNotNil(phoneNumber4)
-            let phoneNumber5 = try PhoneNumber(rawNumber: "900 253 0000")
-            XCTAssertNotNil(phoneNumber5)
-            let phoneNumber6 = try PhoneNumber(rawNumber: "+1 900 253 0000")
-            XCTAssertNotNil(phoneNumber6)
-            let phoneNumber7 = try PhoneNumber(rawNumber: "+1 900 253 0000")
-            XCTAssertNotNil(phoneNumber7)
-        }
-        catch {
-            XCTFail()
-        }
-    }
-    
-    func testBSNumber() {
-        do {
-            let phoneNumber1 = try PhoneNumber(rawNumber: "242 365 1234", region: "BS")
-            XCTAssertNotNil(phoneNumber1)
-            let phoneNumber2 = try PhoneNumber(rawNumber: "+1 242 365 1234", region: "BS")
-            XCTAssertNotNil(phoneNumber2)
-        }
-        catch {
-            XCTFail()
-        }
-    }
-    
-    func testGBNumber() {
-        do {
-            let phoneNumber1 = try PhoneNumber(rawNumber: "(020) 7031 3000", region: "GB")
-            XCTAssertNotNil(phoneNumber1)
-            let phoneNumber2 = try PhoneNumber(rawNumber: "+44 20 7031 3000", region: "GB")
-            XCTAssertNotNil(phoneNumber2)
-            let phoneNumber3 = try PhoneNumber(rawNumber: "(07912) 345 678", region: "GB")
-            XCTAssertNotNil(phoneNumber3)
-            let phoneNumber4 = try PhoneNumber(rawNumber: "+44 7912 345 678", region: "GB")
-            XCTAssertNotNil(phoneNumber4)
-        }
-        catch {
-            XCTFail()
-        }
-    }
-
-    func testDENumber() {
-        do {
-            let phoneNumber1 = try PhoneNumber(rawNumber: "0291 12345678", region: "DE")
-            XCTAssertNotNil(phoneNumber1)
-            let phoneNumber2 = try PhoneNumber(rawNumber: "+49 291 12345678", region: "DE")
-            XCTAssertNotNil(phoneNumber2)
-            let phoneNumber3 = try PhoneNumber(rawNumber: "04134 1234", region: "DE")
-            XCTAssertNotNil(phoneNumber3)
-            let phoneNumber4 = try PhoneNumber(rawNumber: "09123 12345", region: "DE")
-            XCTAssertNotNil(phoneNumber4)
-            let phoneNumber5 = try PhoneNumber(rawNumber: "+49 8021 2345", region: "DE")
-            XCTAssertNotNil(phoneNumber5)
-        }
-        catch {
-            XCTFail()
-        }
-    }
-    
-    func testITNumber() {
-        do {
-            let phoneNumber1 = try PhoneNumber(rawNumber: "02 3661 8300", region: "IT")
-            XCTAssertNotNil(phoneNumber1)
-            let phoneNumber2 = try PhoneNumber(rawNumber: "+39 02 3661 8300", region: "IT")
-            XCTAssertNotNil(phoneNumber2)
-            let phoneNumber3 = try PhoneNumber(rawNumber: "+390236618300", region: "IT")
-            XCTAssertNotNil(phoneNumber3)
-            let phoneNumber4 = try PhoneNumber(rawNumber: "345 678 901", region: "IT")
-            XCTAssertNotNil(phoneNumber4)
-            let phoneNumber5 = try PhoneNumber(rawNumber: "+39 345 678 901", region: "IT")
-            XCTAssertNotNil(phoneNumber5)
-            let phoneNumber6 = try PhoneNumber(rawNumber: "+39345678901", region: "IT")
-            XCTAssertNotNil(phoneNumber6)
-        }
-        catch {
-            XCTFail()
-        }
-    }
-    
-    func testAUNumber() {
-        do {
-            let phoneNumber1 = try PhoneNumber(rawNumber: "02 3661 8300", region: "AU")
-            XCTAssertNotNil(phoneNumber1)
-            let phoneNumber2 = try PhoneNumber(rawNumber: "+61 2 3661 8300", region: "AU")
-            XCTAssertNotNil(phoneNumber2)
-            let phoneNumber3 = try PhoneNumber(rawNumber: "+61236618300", region: "AU")
-            XCTAssertNotNil(phoneNumber3)
-            let phoneNumber4 = try PhoneNumber(rawNumber: "1800 123 456", region: "AU")
-            XCTAssertNotNil(phoneNumber4)
-            let phoneNumber5 = try PhoneNumber(rawNumber: "+61 1800 123 456", region: "AU")
-            XCTAssertNotNil(phoneNumber5)
-            let phoneNumber6 = try PhoneNumber(rawNumber: "+611800123456", region: "AU")
-            XCTAssertNotNil(phoneNumber6)
-        }
-        catch {
-            XCTFail()
-        }
-    }
-//
-    func testARNumber() {
-        do {
-            let phoneNumber1 = try PhoneNumber(rawNumber: "011 8765-4321", region: "AR")
-            XCTAssertNotNil(phoneNumber1)
-            let phoneNumber2 = try PhoneNumber(rawNumber: "+54 11 8765-4321", region: "AR")
-            XCTAssertNotNil(phoneNumber2)
-            let phoneNumber3 = try PhoneNumber(rawNumber: "+541187654321", region: "AR")
-            XCTAssertNotNil(phoneNumber3)
-            let phoneNumber4 = try PhoneNumber(rawNumber: "011 15 8765-4321", region: "AR")
-            XCTAssertNotNil(phoneNumber4)
-            let phoneNumber5 = try PhoneNumber(rawNumber: "+54 9 11 8765 4321", region: "AR")
-            XCTAssertNotNil(phoneNumber5)
-            let phoneNumber6 = try PhoneNumber(rawNumber: "+5491187654321", region: "AR")
-            XCTAssertNotNil(phoneNumber6)
-        }
-        catch {
-            XCTFail()
-        }
-    }
-
-    func testMXNumber() {
-        do {
-            let phoneNumber1 = try PhoneNumber(rawNumber: "045 234 567 8900", region: "MX")
-            XCTAssertNotNil(phoneNumber1)
-            let phoneNumber2 = try PhoneNumber(rawNumber: "+52 1 234 567 8900", region: "MX")
-            XCTAssertNotNil(phoneNumber2)
-            let phoneNumber3 = try PhoneNumber(rawNumber: "+5212345678900", region: "MX")
-            XCTAssertNotNil(phoneNumber3)
-            let phoneNumber5 = try PhoneNumber(rawNumber: "045 55 1234 5678", region: "MX")
-            XCTAssertNotNil(phoneNumber5)
-            let phoneNumber6 = try PhoneNumber(rawNumber: "+52 1 55 1234 5678", region: "MX")
-            XCTAssertNotNil(phoneNumber6)
-        }
-        catch {
-            XCTFail()
-        }
-    }
-//
-//
-
-//
-//    #pragma mark - testFormatMXNumber
-//    {
-//    XCTAssertEqualObjects(@"045 234 567 8900", [_aUtil format:MX_MOBILE1 numberFormat:NBEPhoneNumberFormatNATIONAL]);
-//    XCTAssertEqualObjects(@"+52 1 234 567 8900", [_aUtil format:MX_MOBILE1 numberFormat:NBEPhoneNumberFormatINTERNATIONAL]);
-//    XCTAssertEqualObjects(@"+5212345678900", [_aUtil format:MX_MOBILE1 numberFormat:NBEPhoneNumberFormatE164]);
-//    XCTAssertEqualObjects(@"045 55 1234 5678", [_aUtil format:MX_MOBILE2 numberFormat:NBEPhoneNumberFormatNATIONAL]);
-//    XCTAssertEqualObjects(@"+52 1 55 1234 5678", [_aUtil format:MX_MOBILE2 numberFormat:NBEPhoneNumberFormatINTERNATIONAL]);
-//    XCTAssertEqualObjects(@"+5215512345678", [_aUtil format:MX_MOBILE2 numberFormat:NBEPhoneNumberFormatE164]);
-//    XCTAssertEqualObjects(@"01 33 1234 5678", [_aUtil format:MX_NUMBER1 numberFormat:NBEPhoneNumberFormatNATIONAL]);
-//    XCTAssertEqualObjects(@"+52 33 1234 5678", [_aUtil format:MX_NUMBER1 numberFormat:NBEPhoneNumberFormatINTERNATIONAL]);
-//    XCTAssertEqualObjects(@"+523312345678", [_aUtil format:MX_NUMBER1 numberFormat:NBEPhoneNumberFormatE164]);
-//    XCTAssertEqualObjects(@"01 821 123 4567", [_aUtil format:MX_NUMBER2 numberFormat:NBEPhoneNumberFormatNATIONAL]);
-//    XCTAssertEqualObjects(@"+52 821 123 4567", [_aUtil format:MX_NUMBER2 numberFormat:NBEPhoneNumberFormatINTERNATIONAL]);
-//    XCTAssertEqualObjects(@"+528211234567", [_aUtil format:MX_NUMBER2 numberFormat:NBEPhoneNumberFormatE164]);
-//    }
-//    
-//    #pragma mark - testFormatOutOfCountryCallingNumber
-//    {
-//    XCTAssertEqualObjects(@"00 1 900 253 0000", [_aUtil formatOutOfCountryCallingNumber:US_PREMIUM regionCallingFrom:@"DE"]);
-//    XCTAssertEqualObjects(@"1 650 253 0000", [_aUtil formatOutOfCountryCallingNumber:US_NUMBER regionCallingFrom:@"BS"]);
-//    XCTAssertEqualObjects(@"00 1 650 253 0000", [_aUtil formatOutOfCountryCallingNumber:US_NUMBER regionCallingFrom:@"PL"]);
-//    XCTAssertEqualObjects(@"011 44 7912 345 678", [_aUtil formatOutOfCountryCallingNumber:GB_MOBILE regionCallingFrom:@"US"]);
-//    XCTAssertEqualObjects(@"00 49 1234", [_aUtil formatOutOfCountryCallingNumber:DE_SHORT_NUMBER regionCallingFrom:@"GB"]);
-//    // Note this number is correctly formatted without national prefix. Most of
-//    // the numbers that are treated as invalid numbers by the library are short
-//    // numbers, and they are usually not dialed with national prefix.
-//    XCTAssertEqualObjects(@"1234", [_aUtil formatOutOfCountryCallingNumber:DE_SHORT_NUMBER regionCallingFrom:@"DE"]);
-//    XCTAssertEqualObjects(@"011 39 02 3661 8300", [_aUtil formatOutOfCountryCallingNumber:IT_NUMBER regionCallingFrom:@"US"]);
-//    XCTAssertEqualObjects(@"02 3661 8300", [_aUtil formatOutOfCountryCallingNumber:IT_NUMBER regionCallingFrom:@"IT"]);
-//    XCTAssertEqualObjects(@"+39 02 3661 8300", [_aUtil formatOutOfCountryCallingNumber:IT_NUMBER regionCallingFrom:@"SG"]);
-//    XCTAssertEqualObjects(@"6521 8000", [_aUtil formatOutOfCountryCallingNumber:SG_NUMBER regionCallingFrom:@"SG"]);
-//    XCTAssertEqualObjects(@"011 54 9 11 8765 4321", [_aUtil formatOutOfCountryCallingNumber:AR_MOBILE regionCallingFrom:@"US"]);
-//    XCTAssertEqualObjects(@"011 800 1234 5678", [_aUtil formatOutOfCountryCallingNumber:INTERNATIONAL_TOLL_FREE regionCallingFrom:@"US"]);
-//    
-//    id arNumberWithExtn = [AR_MOBILE copy];
-//    [arNumberWithExtn setExtension:@"1234"];
-//    XCTAssertEqualObjects(@"011 54 9 11 8765 4321 ext. 1234", [_aUtil formatOutOfCountryCallingNumber:arNumberWithExtn regionCallingFrom:@"US"]);
-//    XCTAssertEqualObjects(@"0011 54 9 11 8765 4321 ext. 1234", [_aUtil formatOutOfCountryCallingNumber:arNumberWithExtn regionCallingFrom:@"AU"]);
-//    XCTAssertEqualObjects(@"011 15 8765-4321 ext. 1234", [_aUtil formatOutOfCountryCallingNumber:arNumberWithExtn regionCallingFrom:@"AR"]);
-//    }
-//    
-//    
-//    #pragma mark - testFormatOutOfCountryWithInvalidRegion
-//    {
-//    // AQ/Antarctica isn't a valid region code for phone number formatting,
-//    // so this falls back to intl formatting.
-//    XCTAssertEqualObjects(@"+1 650 253 0000", [_aUtil formatOutOfCountryCallingNumber:US_NUMBER regionCallingFrom:@"AQ"]);
-//    // For region code 001, the out-of-country format always turns into the
-//    // international format.
-//    XCTAssertEqualObjects(@"+1 650 253 0000", [_aUtil formatOutOfCountryCallingNumber:US_NUMBER regionCallingFrom:@"001"]);
-//    }
-//    
-//    
-//    #pragma mark - testFormatOutOfCountryWithPreferredIntlPrefix
-//    {
-//    // This should use 0011, since that is the preferred international prefix
-//    // (both 0011 and 0012 are accepted as possible international prefixes in our
-//    // test metadta.)
-//    XCTAssertEqualObjects(@"0011 39 02 3661 8300", [_aUtil formatOutOfCountryCallingNumber:IT_NUMBER regionCallingFrom:@"AU"]);
-//    }
-//    
-//    
-//    #pragma mark - testFormatOutOfCountryKeepingAlphaChars
-//    {
-//    id alphaNumericNumber = [[NBPhoneNumber alloc] init];
-//    [alphaNumericNumber setCountryCode:@1];
-//    [alphaNumericNumber setNationalNumber:@8007493524];
-//    [alphaNumericNumber setRawInput:@"1800 six-flag"];
-//    XCTAssertEqualObjects(@"0011 1 800 SIX-FLAG", [_aUtil formatOutOfCountryKeepingAlphaChars:alphaNumericNumber regionCallingFrom:@"AU"]);
-//    
-//    [alphaNumericNumber setRawInput:@"1-800-SIX-flag"];
-//    XCTAssertEqualObjects(@"0011 1 800-SIX-FLAG", [_aUtil formatOutOfCountryKeepingAlphaChars:alphaNumericNumber regionCallingFrom:@"AU"]);
-//    
-//    [alphaNumericNumber setRawInput:@"Call us from UK: 00 1 800 SIX-flag"];
-//    XCTAssertEqualObjects(@"0011 1 800 SIX-FLAG", [_aUtil formatOutOfCountryKeepingAlphaChars:alphaNumericNumber regionCallingFrom:@"AU"]);
-//    
-//    [alphaNumericNumber setRawInput:@"800 SIX-flag"];
-//    XCTAssertEqualObjects(@"0011 1 800 SIX-FLAG", [_aUtil formatOutOfCountryKeepingAlphaChars:alphaNumericNumber regionCallingFrom:@"AU"]);
-//    
-//    // Formatting from within the NANPA region.
-//    XCTAssertEqualObjects(@"1 800 SIX-FLAG", [_aUtil formatOutOfCountryKeepingAlphaChars:alphaNumericNumber regionCallingFrom:@"US"]);
-//    XCTAssertEqualObjects(@"1 800 SIX-FLAG", [_aUtil formatOutOfCountryKeepingAlphaChars:alphaNumericNumber regionCallingFrom:@"BS"]);
-//    
-//    // Testing that if the raw input doesn't exist, it is formatted using
-//    // formatOutOfCountryCallingNumber.
-//    [alphaNumericNumber setRawInput:nil];
-//    XCTAssertEqualObjects(@"00 1 800 749 3524", [_aUtil formatOutOfCountryKeepingAlphaChars:alphaNumericNumber regionCallingFrom:@"DE"]);
-//    
-//    // Testing AU alpha number formatted from Australia.
-//    [alphaNumericNumber setCountryCode:@61];
-//    [alphaNumericNumber setNationalNumber:@827493524];
-//    [alphaNumericNumber setRawInput:@"+61 82749-FLAG"];
-//    // This number should have the national prefix fixed.
-//    XCTAssertEqualObjects(@"082749-FLAG", [_aUtil formatOutOfCountryKeepingAlphaChars:alphaNumericNumber regionCallingFrom:@"AU"]);
-//    
-//    [alphaNumericNumber setRawInput:@"082749-FLAG"];
-//    XCTAssertEqualObjects(@"082749-FLAG", [_aUtil formatOutOfCountryKeepingAlphaChars:alphaNumericNumber regionCallingFrom:@"AU"]);
-//    
-//    [alphaNumericNumber setNationalNumber:@18007493524];
-//    [alphaNumericNumber setRawInput:@"1-800-SIX-flag"];
-//    // This number should not have the national prefix prefixed, in accordance
-//    // with the override for this specific formatting rule.
-//    XCTAssertEqualObjects(@"1-800-SIX-FLAG", [_aUtil formatOutOfCountryKeepingAlphaChars:alphaNumericNumber regionCallingFrom:@"AU"]);
-//    
-//    // The metadata should not be permanently changed, since we copied it before
-//    // modifying patterns. Here we check this.
-//    [alphaNumericNumber setNationalNumber:@1800749352];
-//    XCTAssertEqualObjects(@"1800 749 352", [_aUtil formatOutOfCountryCallingNumber:alphaNumericNumber regionCallingFrom:@"AU"]);
-//    
-//    // Testing a region with multiple international prefixes.
-//    XCTAssertEqualObjects(@"+61 1-800-SIX-FLAG", [_aUtil formatOutOfCountryKeepingAlphaChars:alphaNumericNumber regionCallingFrom:@"SG"]);
-//    // Testing the case of calling from a non-supported region.
-//    XCTAssertEqualObjects(@"+61 1-800-SIX-FLAG", [_aUtil formatOutOfCountryKeepingAlphaChars:alphaNumericNumber regionCallingFrom:@"AQ"]);
-//    
-//    // Testing the case with an invalid country calling code.
-//    [alphaNumericNumber setCountryCode:0];
-//    [alphaNumericNumber setNationalNumber:@18007493524];
-//    [alphaNumericNumber setRawInput:@"1-800-SIX-flag"];
-//    // Uses the raw input only.
-//    XCTAssertEqualObjects(@"1-800-SIX-flag", [_aUtil formatOutOfCountryKeepingAlphaChars:alphaNumericNumber regionCallingFrom:@"DE"]);
-//    
-//    // Testing the case of an invalid alpha number.
-//    [alphaNumericNumber setCountryCode:@1];
-//    [alphaNumericNumber setNationalNumber:@80749];
-//    [alphaNumericNumber setRawInput:@"180-SIX"];
-//    // No country-code stripping can be done.
-//    XCTAssertEqualObjects(@"00 1 180-SIX", [_aUtil formatOutOfCountryKeepingAlphaChars:alphaNumericNumber regionCallingFrom:@"DE"]);
-//    
-//    // Testing the case of calling from a non-supported region.
-//    [alphaNumericNumber setCountryCode:@1];
-//    [alphaNumericNumber setNationalNumber:@80749];
-//    [alphaNumericNumber setRawInput:@"180-SIX"];
-//    // No country-code stripping can be done since the number is invalid.
-//    XCTAssertEqualObjects(@"+1 180-SIX", [_aUtil formatOutOfCountryKeepingAlphaChars:alphaNumericNumber regionCallingFrom:@"AQ"]);
-//    }
-    
-
 
 }
